miApiJWT/requirements.txt
fastapi
uvicorn[standard]
email-validator
python-jose[cryptography]
passlib[bcrypt]
python-multipart